requirements.txt
gradio
google-cloud-aiplatform
python-dotenv 
google-auth 
google-auth-oauthlib 
google-auth-httplib2